% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\name{au.table}
\alias{au.table}
\title{Import the Author Table}
\usage{
au.table(ver = "upd")
}
\arguments{
\item{ver}{Version of the database. Default is upd, for the latest version.}
}
\value{
A data frame containing a list of Authors whose work has been included the database.
}
\description{
Imports the Author table from the database.
}
\examples{
au.table()
}
