% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\name{chem.table}
\alias{chem.table}
\title{Import Chemical Table}
\usage{
chem.table(query = NA, ver = "upd")
}
\arguments{
\item{query}{Group name, either a single input or a list. Default is NA.}

\item{ver}{Version of the database. Default is upd, for the latest version.}
}
\value{
A data frame containing identifying data for chemicals in the database.
}
\description{
Imports the chemical table from the database. The table can be subset to include general categories of chemicals.
}
\examples{
chem.table()
chem.table("PCB")
}
