% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get.categories}
\alias{get.categories}
\title{Get a list of Chemical Categories}
\usage{
get.categories(ver = "upd")
}
\arguments{
\item{ver}{Version of the database. Default is upd, for the latest version.}
}
\value{
A data frame containing the categories for all chemicals included the database.
}
\description{
Creates a list of chemical categories included in the database. Note this is not comprehensive.
}
\examples{
get.categories()

}
