% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get.citations}
\alias{get.citations}
\title{Get a list of Citations}
\usage{
get.citations(ver = "upd")
}
\arguments{
\item{ver}{Version of the database. Default is upd, for the latest version.}
}
\value{
A data frame containing citations for all references included the database.
}
\description{
Creates a list of citations included in the database.
}
\examples{
get.citations()

}
