% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{get.groups}
\alias{get.groups}
\title{Get a list of Groups}
\usage{
get.groups(ver = "upd")
}
\arguments{
\item{ver}{Version of the database. Default is upd, for the latest version.}
}
\value{
A data frame containing research groups included the database.
}
\description{
Creates a list of research groups included in the database.
}
\examples{
get.groups()

}
