% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\name{koa.table}
\alias{koa.table}
\title{Import the KOA Table}
\usage{
koa.table(ver = "upd")
}
\arguments{
\item{ver}{Version of the database. Default is upd, for the latest version.}
}
\value{
A data frame containing KOA data included the database.
}
\description{
Imports the KOA table from the database.
}
\examples{
koa.table()
koa.table(upd)
}
