% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchdatabase.R
\name{open.database}
\alias{open.database}
\title{Open the KOA database}
\usage{
\method{open}{database}(ver = "upd")
}
\arguments{
\item{ver}{Version of the database. Default is upd, for the latest version.}
}
\value{
A connection to the KOA database.
}
\description{
Opens the KOA database using RSQLite.
}
\examples{
open.database()
open.database("upd")
}
