% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\name{prop.table}
\alias{prop.table}
\title{Import the Property Table}
\usage{
prop.table(query = NA, ver = "upd")
}
\arguments{
\item{query}{Property type, either a single input or a list. Default is NA. Must be either: "log_KOW", "log_KAW", "Sw", "VP", "inf_act", "dG".}

\item{ver}{Version of the database. Default is upd, for the latest version.}
}
\value{
A data frame containing information on the different methods used to obtain the KOA values in the database.
}
\description{
Imports the Property table from the database.
}
\examples{
prop.table()
prop.table("Dynamic")
}
