% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchdatabase.R
\name{query.category}
\alias{query.category}
\title{Search by chemical category}
\usage{
query.category(query, more.info = F, ver = "upd")
}
\arguments{
\item{query}{Category name, either a single input or a vector.}

\item{more.info}{When FALSE (default), returns fewer columns from the chemical, methods, and reference tables. When TRUE, additional details on the chemical (category, synonyms, molar mass), methods (wet/dry octanol, type of value reported, type of method), and reference (group/team).}

\item{ver}{Version of the database. Default is upd, for the latest version.}
}
\value{
A data frame containing KOA data for chemicals which belong to the category or categories matching the query.
}
\description{
Create a data frame with all estimated and experimental KOA values for chemicals which belong to the specified category. All entries are case sensitive.
}
\examples{
query.category("PCB")
query.category(c("Terpene", "PBDE"))
query.category(c("Terpene", "PBDE"), more.info = TRUE)
query.category("PCB", ver = "upd")
}
\seealso{
Other Queries: 
\code{\link{query.cas}()},
\code{\link{query.citation}()},
\code{\link{query.group}()},
\code{\link{query.mass}()},
\code{\link{query.meth.type}()},
\code{\link{query.meth}()},
\code{\link{query.name}()}
}
\concept{Queries}
