% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchdatabase.R
\name{query.citation}
\alias{query.citation}
\title{Search by citation}
\usage{
query.citation(query, more.info = F, ver = "upd")
}
\arguments{
\item{query}{Citation, either a single input or a vector. Use function ref.table() to get a list of citations in the database.}

\item{more.info}{When FALSE (default), returns fewer columns from the chemical, methods, and reference tables. When TRUE, additional details on the chemical (category, synonyms, molar mass), methods (wet/dry octanol, type of value reported, type of method), and reference (group/team).}

\item{ver}{Version of the database. Default is upd, for the latest version.}
}
\value{
A data frame containing KOA data from the citations included in the query.
}
\description{
Create a data frame with all estimated and experimental KOA values from the citations in the query.
}
\examples{
query.citation("Lei et al. 2019")
query.citation(c("Hussam and Carr 1985", "Lei et al. 2019"))
query.citation("Lei et al. 2019", more.info = TRUE, ver = "upd")
}
\seealso{
Other Queries: 
\code{\link{query.cas}()},
\code{\link{query.category}()},
\code{\link{query.group}()},
\code{\link{query.mass}()},
\code{\link{query.meth.type}()},
\code{\link{query.meth}()},
\code{\link{query.name}()}
}
\concept{Queries}
