% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchdatabase.R
\name{query.group}
\alias{query.group}
\title{Search by group}
\usage{
query.group(query, more.info = F, ver = "upd")
}
\arguments{
\item{query}{Group name, either a single input or a vector.}

\item{more.info}{When FALSE (default), returns fewer columns from the chemical, methods, and reference tables. When TRUE, additional details on the chemical (category, synonyms, molar mass), methods (wet/dry octanol, type of value reported, type of method), and reference (group/team).}

\item{ver}{Version of the database. Default is upd, for the latest version.}
}
\value{
A data frame containing KOA data from the groups included in the query.
}
\description{
Create a data frame with all estimated and experimental KOA values from the research groups in the query.
}
\examples{
query.group("Harner")
query.group(c("Harner", "Chen", "Odabasi"))
query.group(c("Harner", "Chen", "Odabasi"), more.info = TRUE)
query.group("Harner", ver = "upd")
}
