% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchdatabase.R
\name{query.mass}
\alias{query.mass}
\title{Search by chemical mass}
\usage{
query.mass(lower_limit, upper_limit, ver = "upd")
}
\arguments{
\item{lower_limit}{Lower mass limit.}

\item{upper_limit}{Upper mass limit.}

\item{ver}{Version of the database. Default is upd, for the latest version.}

\item{more.info}{When FALSE (default), returns fewer columns from the chemical, methods, and reference tables. When TRUE, additional details on the chemical (category, synonyms, molar mass), methods (wet/dry octanol, type of value reported, type of method), and reference (group/team).}
}
\value{
A data frame containing KOA data for chemicals which have a mass between the range specified in the query.
}
\description{
Create a data frame with all estimated and experimental KOA values for chemicals which have a molar mass between the two values in the query.
}
\examples{
query.mass(350, 400)
query.mass(200, 350, more.info = TRUE)
query.mass(350, 400, ver = "upd")
}
\seealso{
Other Queries: 
\code{\link{query.cas}()},
\code{\link{query.category}()},
\code{\link{query.citation}()},
\code{\link{query.group}()},
\code{\link{query.name}()}
}
\concept{Queries}
