% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchdatabase.R
\name{query.meth}
\alias{query.meth}
\title{Search by method}
\usage{
query.meth(query, more.info = F, ver = "upd")
}
\arguments{
\item{query}{The specific method.}

\item{more.info}{When FALSE (default), returns fewer columns from the chemical, methods, and reference tables. When TRUE, additional details on the chemical (category, synonyms, molar mass), methods (wet/dry octanol, type of value reported, type of method), and reference (group/team).}

\item{ver}{Version of the database. Default is upd, for the latest version.}
}
\value{
A data frame containing *K*~OA~ data for chemicals with CAS No. matching the query.
}
\description{
Create a data frame with all *K*~OA~ values obtained using a specific method. Method types include: "FM", "QSPR", "Dy-GLC-RT", "HS", "RT", "droplet", "Eqbm", "Solvation", "VP", "GasSol", or "GS".
}
\examples{
query.meth("FM")
query.meth("Eqbm", more.info = TRUE)
}
\seealso{
Other Queries: 
\code{\link{query.cas}()},
\code{\link{query.category}()},
\code{\link{query.citation}()},
\code{\link{query.group}()},
\code{\link{query.mass}()},
\code{\link{query.meth.type}()},
\code{\link{query.name}()}
}
\concept{Queries}
