% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchdatabase.R
\name{query.meth.type}
\alias{query.meth.type}
\title{Search by method type}
\usage{
query.meth.type(query, more.info = F, ver = "upd")
}
\arguments{
\item{query}{The type of method.}

\item{more.info}{When FALSE (default), returns fewer columns from the chemical, methods, and reference tables. When TRUE, additional details on the chemical (category, synonyms, molar mass), methods (wet/dry octanol, type of value reported, type of method), and reference (group/team).}

\item{ver}{Version of the database. Default is upd, for the latest version.}
}
\value{
A data frame containing KOA data for chemicals with CAS No. matching the query.
}
\description{
Create a data frame with all KOA values obtained using a specific type of method. Method types include: "Dynamic", "Static", "Indirect", or "EST".
}
\examples{
query.meth.type("Dynamic")
query.meth.type("Dynamic", more.info = TRUE)
}
\seealso{
Other Queries: 
\code{\link{query.cas}()},
\code{\link{query.category}()},
\code{\link{query.citation}()},
\code{\link{query.group}()},
\code{\link{query.mass}()},
\code{\link{query.meth}()},
\code{\link{query.name}()}
}
\concept{Queries}
