% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchdatabase.R
\name{query.name}
\alias{query.name}
\title{Search by chemical name}
\usage{
query.name(query, more.info = F, ver = "upd")
}
\arguments{
\item{query}{Chemical name, either a single input or a vector. Searches against multiple synonymns included in the chemical table.}

\item{more.info}{When FALSE (default), returns fewer columns from the chemical, methods, and reference tables. When TRUE, additional details on the chemical (category, synonyms, molar mass), methods (wet/dry octanol, type of value reported, type of method), and reference (group/team).}

\item{ver}{Version of the database. Default is upd, for the latest version.}
}
\value{
A data frame containing KOA data for chemicals with names exactly matching the query.
}
\description{
Create a data frame with all estimated and experimental KOA values which exactly match the specified chemical name(s). All entries are case sensitive.
}
\examples{
query.name("8:2 FTOH")
query.name("HCB", more.info = TRUE)
query.name(c("8:2 FTOH", "Hexachlorobenzene"))
query.name("8:2 FTOH", ver = "upd")
}
