% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\name{table.meth}
\alias{table.meth}
\title{Import the Methods Table}
\usage{
table.meth(query = NA, ver = "upd")
}
\arguments{
\item{query}{Method Type, either a single input or a list. Default is NA. Must be either: Dynamic, Static, Indirect, or EST.}

\item{ver}{Version of the database. Default is upd, for the latest version.}
}
\value{
A data frame containing information on the different methods used to obtain the KOA values in the database.
}
\description{
Imports the Methods table from the database.
}
\examples{
table.meth()
table.meth("Dynamic")
}
