% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\name{table.ref}
\alias{table.ref}
\title{Import the Reference Table}
\usage{
table.ref(query = NA, ver = "upd")
}
\arguments{
\item{query}{Group name, either a single input or a list. Default is NA.}

\item{ver}{Version of the database. Default is upd, for the latest version.}
}
\value{
A data frame containing KOA data included the database.
}
\description{
Imports the Reference table from the database.
}
\examples{
table.ref()
table.ref("Harner")
table.ref("Carr", "upd")

}
