% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchdatabase.R
\name{koa.database}
\alias{koa.database}
\title{Open the *K*~OA~ database}
\usage{
koa.database(ver = "upd")
}
\arguments{
\item{ver}{Version of the database. Default is upd, for the latest version.}
}
\value{
A connection to the *K*~OA~ database.
}
\description{
Opens the *K*~OA~ database using RSQLite. The database connection must be closed before other functions in the package can be used.
}
\examples{
conn <- koa.database("upd")
DBI::dbDisconnect(conn)

}
